% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{block_insert_functions}
\alias{block_insert_functions}
\title{determine which function to use for inserting macros into blocks}
\usage{
block_insert_functions(scaffold, block_names)
}
\arguments{
\item{scaffold}{`stan_blocks` of scaffold}

\item{block_names}{names of blocks to insert}
}
\description{
determine which function to use for inserting macros into blocks
}
