% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{blocks_to_stan_string}
\alias{blocks_to_stan_string}
\title{convert a block list into stan code; opposite of `get_blocks`}
\usage{
blocks_to_stan_string(block_list)
}
\arguments{
\item{block_list}{a `stan_blocks` object}
}
\value{
text of a stan file
}
\description{
convert a block list into stan code; opposite of `get_blocks`
}
