% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_stan_macro.r
\name{define_stan_macro}
\alias{define_stan_macro}
\title{define_stan_macro: create a function that makes stan macros convienient}
\usage{
define_stan_macro(.args = alist(), ..., .glue_control = list(.open =
  "{{", .close = "}}"))
}
\arguments{
\item{.args}{alist of argument names that must be subbed}

\item{...}{named character vectors defining a section of the macro; these will be used with `glue`}

\item{.glue_control}{named list of control args for glue}
}
\value{
a function that with arguments .section, ..., and whatever is in .args
}
\description{
define_stan_macro: create a function that makes stan macros convienient
}
