% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_util.r
\name{extract_all_delims_linear}
\alias{extract_all_delims_linear}
\title{This pulls out all text groups surrounded by the delims, from left to right}
\usage{
extract_all_delims_linear(x, .left, .right)
}
\description{
This pulls out all text groups surrounded by the delims, from left to right
}
