% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_util.r
\name{extract_delim_linear}
\alias{extract_delim_linear}
\title{this searches for left and right delimiters, but in a non-nested sense
The next one it finds gets triggered}
\usage{
extract_delim_linear(x, .left, .right)
}
\description{
this searches for left and right delimiters, but in a non-nested sense
The next one it finds gets triggered
}
