% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_util.r
\name{extract_sequence}
\alias{extract_sequence}
\title{This removes a number of delimiter-based sequences
.lefts and .rights much match in length
there are removed sequentially, by delimiter then from right to left}
\usage{
extract_sequence(x, .lefts = c("//", "/*"), .rights = c("\\n", "*/"))
}
\description{
This removes a number of delimiter-based sequences
.lefts and .rights much match in length
there are removed sequentially, by delimiter then from right to left
}
