% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_util.r
\name{filter_rows}
\alias{filter_rows}
\title{Basically dplyr::filter with one condition}
\usage{
filter_rows(.df, .cond)
}
\arguments{
\item{.df}{a data frame}

\item{.cond}{a logical statement about .df}
}
\description{
Basically dplyr::filter with one condition
}
