% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_file_util.r
\name{find_block}
\alias{find_block}
\title{parse a stan file and pull out the corresponding block
note that two extra block names are here}
\usage{
find_block(x, block = c("macro args", "use macros", "functions", "data",
  "transformed data", "parameters", "transformed parameters", "model",
  "generated quantities"), ...)
}
\description{
parse a stan file and pull out the corresponding block
note that two extra block names are here
}
