% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_macro.r
\name{get_arg_names}
\alias{get_arg_names}
\title{gets the names of parsed macro arguments}
\usage{
get_arg_names(.args)
}
\arguments{
\item{.args}{list of language objects returned by `get_macro_args`}
}
\value{
a character vector of argument names
}
\description{
gets the names of parsed macro arguments
}
