% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_file_util.r
\name{get_blocks}
\alias{get_blocks}
\title{transforms a stan file into list with named elements corresponding to the program block}
\usage{
get_blocks(x)
}
\arguments{
\item{x}{stan code}
}
\description{
transforms a stan file into list with named elements corresponding to the program block
}
