% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_util.r
\name{get_delim_contents}
\alias{get_delim_contents}
\title{returns what's inside the first full pair of delimeters}
\usage{
get_delim_contents(x, .left = "{", .right = "}")
}
\description{
returns what's inside the first full pair of delimeters
}
