% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_util.r
\name{get_delim_tbl}
\alias{get_delim_tbl}
\title{this is used to find matched delimiters
it returns the position of each right and left delim,
in order.}
\usage{
get_delim_tbl(x, .left = "{", .right = "}")
}
\description{
this is used to find matched delimiters
it returns the position of each right and left delim,
in order.
}
