% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{get_input_code}
\alias{get_input_code}
\title{Determines whether the input is a file or charcter vector of stan code}
\usage{
get_input_code(input)
}
\arguments{
\item{input}{from parse_stan_macros}
}
\description{
Determines whether the input is a file or charcter vector of stan code
}
