% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_macro.r
\name{get_macro_args}
\alias{get_macro_args}
\title{extracts the arguments from a macro args block}
\usage{
get_macro_args(blocks)
}
\arguments{
\item{blocks}{a macro file parsed with `get_blocks`}
}
\value{
a list of language objects, one per argument
}
\description{
extracts the arguments from a macro args block
}
