% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{get_macro_calls}
\alias{get_macro_calls}
\title{extracts the macro calls from the use macros block}
\usage{
get_macro_calls(blocks)
}
\arguments{
\item{blocks}{a macro file parsed with `get_blocks`}
}
\value{
a list of language objects, one per argument
}
\description{
extracts the macro calls from the use macros block
}
