% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_macro.r
\name{get_macro_file_list}
\alias{get_macro_file_list}
\title{reads in a list of macro files}
\usage{
get_macro_file_list(macro_files)
}
\arguments{
\item{macro_files}{a list of file names, corresponding to macros}
}
\description{
reads in a list of macro files
}
