% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{has_value}
\alias{has_value}
\title{determine if a macro has a value argument}
\usage{
has_value(macro_fun)
}
\arguments{
\item{macro_fun}{object of class `stan_macro`}
}
\description{
determine if a macro has a value argument
}
