% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{insert_macro_section}
\alias{insert_macro_section}
\title{inserts a block of macro(s) into a block of scaffold}
\usage{
insert_macro_section(.scaffold, .macros, pos = NA)
}
\arguments{
\item{.scaffold}{one stan block to insert into}

\item{.macros}{character vector of parsed, merged stan macros for the block}

\item{pos}{where to insert.  If NA, automatically inserts after declarations (if in an appropriate block) or at the end (if not)}
}
\description{
inserts a block of macro(s) into a block of scaffold
}
