% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{insert_macros}
\alias{insert_macros}
\title{insert macros into scaffold}
\usage{
insert_macros(scaffold_list, macro_blocks)
}
\arguments{
\item{scaffold_list}{a `stan_blocks` list to be inserted into}

\item{macro_blocks}{a `stan_blocks` list of parsed macros to insert}
}
\description{
insert macros into scaffold
}
