% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_util.r
\name{is_assignment}
\alias{is_assignment}
\title{Checks to see if an expression is an assignment;}
\usage{
is_assignment(x)
}
\arguments{
\item{x}{a language object to examine}
}
\value{
logical
}
\description{
Checks to see if an expression is an assignment;
}
