% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{macros_exist}
\alias{macros_exist}
\title{This should be a safe way of getting a macro's definition}
\usage{
macros_exist(macro_names, macro_list)
}
\arguments{
\item{macro_names}{character vector}

\item{macro_list}{list of `stan_macro` objects}
}
\description{
This should be a safe way of getting a macro's definition
}
