% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroStan-package.R
\docType{package}
\name{macroStan-package}
\alias{macroStan}
\alias{macroStan-package}
\title{macroStan: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
