% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_util.r
\name{nth}
\alias{nth}
\title{return the index of the first element of x_lgl that is TRUE
returns empty vector if none apply}
\usage{
nth(x_lgl, n = 1)
}
\description{
return the index of the first element of x_lgl that is TRUE
returns empty vector if none apply
}
