% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{order_stan_macros}
\alias{order_stan_macros}
\title{parse the used macros into ordered text,ready for insertion}
\usage{
order_stan_macros(blocks, macro_list = list())
}
\arguments{
\item{blocks}{`get_blocks()` output for the scaffold}

\item{macro_list}{list of available macros}
}
\value{
a named list of blocks
}
\description{
parse the used macros into ordered text,ready for insertion
}
