% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_util.r
\name{parse_assignment}
\alias{parse_assignment}
\title{parse a text assignment, quoting what needs to be quoted}
\usage{
parse_assignment(x, .def_side)
}
\arguments{
\item{x}{text of an assignment}

\item{.def_side}{if there's no assignment, what should it be treated as?}
}
\description{
parse a text assignment, quoting what needs to be quoted
}
