% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_util.r
\name{parse_assignments}
\alias{parse_assignments}
\title{Some stan code won't parse to R.  The only time that it's needed
is when figuring out assignments
This tries to parse to R; if it fails, it wraps the lhs in quotes
and parses that}
\usage{
parse_assignments(x_lst, .def_side = c("lhs", "rhs"))
}
\arguments{
\item{x_lst}{text of an assignment to be parsed}

\item{.def_side}{if there's no assignment, what should it be treated as?}
}
\description{
Some stan code won't parse to R.  The only time that it's needed
is when figuring out assignments
This tries to parse to R; if it fails, it wraps the lhs in quotes
and parses that
}
