% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_macro.r
\name{parse_macro_block}
\alias{parse_macro_block}
\title{generic function used by get_macro_calls and get_macro_args}
\usage{
parse_macro_block(blocks, .def_side = c("rhs", "lhs"),
  block_name = c("use macros", "macro args"))
}
\arguments{
\item{blocks}{a macro file parsed with `get_blocks`}

\item{.def_side}{for `parse_assignments`}

\item{block_name}{"use macros" or "macro args"}
}
\description{
generic function used by get_macro_calls and get_macro_args
}
