% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{parse_macro_calls}
\alias{parse_macro_calls}
\title{identify which macros are being used and parse them into text}
\usage{
parse_macro_calls(scaffold_blocks, macro_list)
}
\arguments{
\item{scaffold_blocks}{`get_blocks()` of the scaffold}

\item{macro_list}{list of available macros}
}
\description{
identify which macros are being used and parse them into text
}
