% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_macro.r
\name{parse_macro_formals}
\alias{parse_macro_formals}
\title{parse the arguments for the R representation of a macro}
\usage{
parse_macro_formals(.args)
}
\arguments{
\item{.args}{list of language objects returned by `get_macro_args`}
}
\value{
formals for a new function
}
\description{
parse the arguments for the R representation of a macro
}
