% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_stan_macros.r
\name{parse_stan_macros}
\alias{parse_stan_macros}
\title{parse_stan_macros: read in a stan file and replace macros}
\usage{
parse_stan_macros(input, out_file = NA, ..., .macro_symbol = "$",
  .macro_list = list(), .glue_control = list())
}
\arguments{
\item{input}{either a file name or a the text of a stan file}

\item{out_file}{output file (can be NA)}

\item{...}{named arguments for string interpolation}

\item{.macro_symbol}{character identifying the beginning of a macro}

\item{.macro_list}{a list of arguments to be used for glue's string interpolation}

\item{.glue_control}{named list of arguments to glue, includeing .open and .close}
}
\value{
the parsed stan code (invisibly if `out_file = NA`)
}
\description{
parse_stan_macros: read in a stan file and replace macros
}
