% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{parse_value}
\alias{parse_value}
\title{converts expression from x = foo(bar = 1) to foo(bar = 1, value = x)
ignores non-assignment expressions}
\usage{
parse_value(.expr)
}
\arguments{
\item{.expr}{language object}
}
\description{
converts expression from x = foo(bar = 1) to foo(bar = 1, value = x)
ignores non-assignment expressions
}
