% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_file_util.r
\name{post_declaration_position}
\alias{post_declaration_position}
\title{determine where declarations inside a block end;
returns the character position immediately after declarations}
\usage{
post_declaration_position(x, .left = "{|")
}
\description{
determine where declarations inside a block end;
returns the character position immediately after declarations
}
