% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_macro.r
\name{read_macro}
\alias{read_macro}
\title{Read in a macro file}
\usage{
read_macro(file, macro_code = readLines(file),
  .glue_control = list(.open = "{|", .close = "|}"))
}
\arguments{
\item{file}{file name to read in; ignored if `macro_code` is used}

\item{macro_code}{text representation of the macro;}

\item{.glue_control}{list of control arguments to pass to `glue::glue_data`}
}
\value{
a stan_macro object, which is a function that returns parsed macro code
}
\description{
Read in a macro file
}
