% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_file_util.r
\name{remove_comments}
\alias{remove_comments}
\title{removes commented parts of a stan block, replacing with the same amount of white space}
\usage{
remove_comments(x)
}
\description{
removes commented parts of a stan block, replacing with the same amount of white space
}
