% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_util.r
\name{sort_by}
\alias{sort_by}
\title{sorts a data frame}
\usage{
sort_by(.df, .expr, .desc = FALSE)
}
\arguments{
\item{.df}{a data frame}

\item{.expr}{what to sort by; can be in terms of columns of `.df`}

\item{.desc}{logical; if true, sorts by descending order (otherwise ascending)}
}
\description{
sorts a data frame
}
