% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_macro.r
\name{stan_macro_generics}
\alias{stan_macro_generics}
\alias{as.list.stan_macro}
\alias{summary.stan_macro}
\alias{print.stan_macro}
\title{Stan macro generic functions}
\usage{
\method{as.list}{stan_macro}(x, ...)

\method{summary}{stan_macro}(object, ...)

\method{print}{stan_macro}(x, ...)
}
\arguments{
\item{x}{stan_macro object}

\item{...}{additional arguments passed to or from other methods}

\item{object}{stan_macro object}
}
\description{
Stan macro generic functions
}
