% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_hs.R
\name{stan_macro_horseshoe}
\alias{stan_macro_horseshoe}
\title{stan_macro_horseshoe: defines components of a regularized horseshoe prior
This should be used with `parse_stan_macros()`}
\usage{
stan_macro_horseshoe(name = 1, N_local = "D", ...,
  .section = sec_names)
}
\arguments{
\item{name}{label for the prior}

\item{N_local}{number of local parameters (character or integer)}

\item{.section}{only return named section; by default, returns all sections}
}
\value{
6 sections
}
\description{
stan_macro_horseshoe: defines components of a regularized horseshoe prior
This should be used with `parse_stan_macros()`
}
