% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{verify_macro_list}
\alias{verify_macro_list}
\title{checks to make sure all provided macros (and files) are of the valid class; returns combined list}
\usage{
verify_macro_list(macro_files, macro_list, ...)
}
\arguments{
\item{macro_files}{named list of macro file names}

\item{macro_list}{named list of macro definitions}

\item{...}{additional named macro definitions}
}
\description{
checks to make sure all provided macros (and files) are of the valid class; returns combined list
}
