% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_macro.r
\name{get_macro_args}
\alias{get_macro_args}
\title{#' extracts the arguments from a macro args block
#' @param blocks a macro file parsed with `get_blocks`
#' @return a list of language objects, one per argument}
\usage{
get_macro_args(blocks)
}
\description{
#' extracts the arguments from a macro args block
#' @param blocks a macro file parsed with `get_blocks`
#' @return a list of language objects, one per argument
}
