% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_stan_macro.r
\name{glue_args}
\alias{glue_args}
\title{alternate paramterization for glue_data, easier to lapply with}
\usage{
glue_args(what, args, control = list(), ...)
}
\description{
alternate paramterization for glue_data, easier to lapply with
}
