% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macro.r
\name{parse_stan_macros}
\alias{parse_stan_macros}
\title{Applies macros to a stan file}
\usage{
parse_stan_macros(input, output_file, macro_list = list(),
  macro_files = list(), ..., .validate_output = TRUE)
}
\arguments{
\item{input}{stan code or file name}

\item{output_file}{file name to output}

\item{macro_list}{named list of macro definitions}

\item{macro_files}{named list of macro file names}

\item{...}{additional named macro definitions}

\item{.validate_output}{check to make sure output stan file is valid}
}
\value{
the parsed stan code if `out_file = NA`; otherwise `out_file`
}
\description{
Applies macros to a stan file
}
